API/requirements.txt
flask
flask-restful
flask-cors
boto3
python-dotenv
torch
opencv-python
Pillow
torchvision